Keep issue 2907 coverage inside CI file budgets
The workflow guard rejected the first PR shape because the new regression tests and routing helpers grew already-budgeted Swift files. Split the issue-specific tests and recoverable route helpers into focused sub-budget files, and keep the TerminalController fallback compact so the guard passes without raising the checked-in budget.

Constraint: workflow-guard-tests enforces the Swift file length budget on PRs

Rejected: Refresh the Swift file length budget | would accept new file-growth debt instead of preserving the guard intent

Confidence: high

Scope-risk: narrow

Directive: Keep future issue-specific socket regressions in focused test files instead of growing WindowAndDragTests.swift

Tested: ./tests/test_ci_swift_file_length_budget.sh

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: xcodebuild test -project GhosttyTabs.xcodeproj -scheme cmux-unit -destination 'platform=macOS' -derivedDataPath /tmp/cmux-issue-2907-ci-fix-test -only-testing:cmuxTests/AppDelegateIssue2907RoutingTests

Tested: ./scripts/reload.sh --tag issue-2907-fix

Not-tested: Full CI rerun after this commit
